+++ pkg/backend/grpcServer/handlers/referencePointCache_test.go
@@ -4,7 +4,7 @@ import (
 	"sync"
 	"testing"
 
-	structs "github.com/Domilz/d7017e-mesh-network/internal/backend/grpcServer/forms"
+	structs "github.com/Domilz/d7017e-mesh-network/pkg/backend/grpcServer/forms"
 	"github.com/stretchr/testify/assert"
 )
