+++ pkg/protocol/utils/insert.go
@@ -1,7 +1,7 @@
 package utils
 
 import (
-	pb "github.com/Domilz/d7017e-mesh-network/internal/protocol/protofiles/tag"
+	pb "github.com/Domilz/d7017e-mesh-network/pkg/protocol/protofiles/tag"
 )
 
 func (stateHandler *StateHandler) InsertMultipleReadings(state *pb.State) {
